♻️ X3NHBJ blueprint: decompose listing commands
