+++ .agentplane/tasks/202605290332-X3NHBJ/pr/github-body.md
@@ -1,0 +1,33 @@
+Task: `202605290332-X3NHBJ`
+Title: Blueprint command listing decomposition
+Canonical task record: `.agentplane/tasks/202605290332-X3NHBJ/README.md`
+
+## Summary
+
+Blueprint command listing decomposition
+
+Decompose packages/agentplane/src/commands/blueprint/blueprint.command.ts by extracting blueprint list/examples command handlers into a focused module, preserving command-loader exports and CLI output while reducing runtime hotspot count.
+
+## Scope
+
+- In scope: Decompose packages/agentplane/src/commands/blueprint/blueprint.command.ts by extracting blueprint list/examples command handlers into a focused module, preserving command-loader exports and CLI output while reducing runtime hotspot count.
+- Out of scope: unrelated refactors not required for "Blueprint command listing decomposition".
+
+## Verification
+
+- State: pending
+- Note: Not recorded yet.
+- Canonical workflow state lives in the task README.
+
+<details>
+<summary>Raw evidence</summary>
+
+- Updated: 2026-05-29T03:33:22.818Z
+- Branch: task/202605290332-X3NHBJ/blueprint-command-listing-decomposition
+- Head: computed live by `agentplane pr check` / `agentplane integrate`
+
+```text
+No changes detected.
+```
+
+</details>
