+++ .agentplane/tasks/202605290332-X3NHBJ/pr/github-title.txt
@@ -1,0 +1,1 @@
+🚧 X3NHBJ task: Blueprint command listing decomposition [202605290332-X3NHBJ]
