+++ .agentplane/tasks/202605290332-X3NHBJ/pr/review.md
@@ -1,0 +1,36 @@
+# PR Review
+
+Created: 2026-05-29T03:33:22.818Z
+
+## Task
+
+- Task: `202605290332-X3NHBJ`
+- Title: Blueprint command listing decomposition
+- Status: DOING
+- Branch: `task/202605290332-X3NHBJ/blueprint-command-listing-decomposition`
+- Canonical task record: `.agentplane/tasks/202605290332-X3NHBJ/README.md`
+
+## Verification
+
+- State: pending
+- Note: Not recorded yet.
+- Canonical workflow state lives in the task README.
+
+## Handoff Notes
+
+- No handoff notes recorded yet. Use `agentplane pr note ...` to append one.
+
+<!-- BEGIN AUTO SUMMARY -->
+<details>
+<summary>Raw evidence</summary>
+
+- Updated: 2026-05-29T03:33:22.818Z
+- Branch: task/202605290332-X3NHBJ/blueprint-command-listing-decomposition
+- Head: computed live by `agentplane pr check` / `agentplane integrate`
+
+```text
+No changes detected.
+```
+
+</details>
+<!-- END AUTO SUMMARY -->
